[base] TimeSpan: fixes rounding issues in fromMilliseconds()

+++ include/x0/TimeSpan.h
@@ -45,7 +45,7 @@ public:
 	static TimeSpan fromHours(std::size_t v) { return TimeSpan(ticksPerHour() * v); }
 	static TimeSpan fromMinutes(std::size_t v) { return TimeSpan(ticksPerMinute() * v); }
 	static TimeSpan fromSeconds(std::size_t v) { return TimeSpan(ticksPerSecond() * v); }
-	static TimeSpan fromMilliseconds(std::size_t v) { return TimeSpan(ev_tstamp(v / 1000 + (v % 1000) / 1000.0f)); }
+	static TimeSpan fromMilliseconds(std::size_t v) { return TimeSpan(ev_tstamp(v / 1000 + (unsigned(v) % 1000) / 1000.0)); }
 
 	std::size_t totalMilliseconds() const { return value_ * 1000; }
